+++ src/pyfreepbx/facade.py
@@ -32,6 +32,7 @@ from pyfreepbx.exceptions import ConfigError
 from pyfreepbx.logging import get_logger
 from pyfreepbx.models.health import StatusResult
 from pyfreepbx.services.extensions import ExtensionService
+from pyfreepbx.services.firewall import FirewallService
 from pyfreepbx.services.health import HealthService
 from pyfreepbx.services.queues import QueueService
 from pyfreepbx.services.system import SystemService
@@ -102,10 +103,11 @@ class FreePBX:
             self._ami_client = AMIClient(self._ami_config)
 
         # Services
-        self._extensions = ExtensionService(self._client)
+        self._extensions = ExtensionService(self._client, self._rest_client)
         self._queues = QueueService(self._client, self._ami_client)
         self._system = SystemService(self._client, self._ami_client)
         self._health = HealthService(self._client, self._ami_client)
+        self._firewall = FirewallService(self._rest_client)
 
         log.debug("FreePBX facade initialized for %s", host)
 
@@ -278,6 +280,10 @@ class FreePBX:
     @property
     def health(self) -> HealthService:
         return self._health
+
+    @property
+    def firewall(self) -> FirewallService:
+        return self._firewall
 
     @property
     def rest(self) -> RestClient: